Notebooks/requirements.txt
pre-commit
black
blacken-docs==0.5.0
pylint==2.3.1
pylint-venv
tensorflow
numpy
scipy
pandas
matplotlib
jupyter
Pillow
requests
pyqt5
